Reconfiguration and new main Page
In this one basically I reconfigured a lot of things including the previous index page new at first I want the user to have a page where they can see all their documents and be able to create new documents. so for that I had to create a new directory called main and the previoius directory in which you could see the document is now named document instead of being global scoped.

I also moved the user.ts to a global scope inside the html dir and made changes to it as well like making a constructors for the document interface new can be used when a new document is created.

I have also switched from having all my scripts dir inside each of the page to new being in a dlobally scoped, this was needed so that ts can be configured easily.