Reconfig of files and new Main page where user can find all docs
I had to do lots of configuration because of typescript and how it liked the project to be structured as which was hard because I wanted to have the user.ts file as global to both the main and document folders could handle them but what ended up happening was I did create the two folders but then I just have all the files in one folder after that what I am planning to do is just use webpack to make all my code one file I looked it up easy enough. For the main page I made a lot of complex types basically allowing the user to be able to change the things the want I will add the functionality as I go though them but the basic groundwork I believe is set up, Now I am going to change course again and look more into how to do the changes in the text in a specific part which after which the basic format could be applied to almost any aspect it would just be a matter of how much time it would take me to type it. But after I am done with that part of the project I will start to take a look at the back end of the project. Which does include a lot of things like some routing, then making a database using PostgreSQL, and after all that I will finally try to deploy it. I think I have a solid week or 2 week worth of work to do. but we are making some progress